reconstruction_stella/requirements.txt
celery
redis
pyyaml
requests
